Task-4(Docker Basics)/requirements.txt
langchain-perplexity
langchain-core
python-dotenv